+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   - Issue loading `package.json` in `setup.py` due to encoding
   - Force UTF-8 encoding on I/O
 
+### Changed
+  - Split the CoNLL-2009/mate format into a Gold and a System version
+
 ## [0.5.1] - 2017-04-03
 [0.5.1]: https://github.com/LoicGrobol/ginger/compare/v0.5.0...v0.5.1
 ### Fixed